Function notation started!
There is much to work on! But this is the basic gist of what we want
right? It is not shareable, it doesn't respond well to updates from the
core function, and for some reason it gives back whatever weird results
when the f(x) is the first thing up. There is something screwy with the
code, but how?